% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{GenerateNoise}
\alias{GenerateNoise}
\title{Noise Generator}
\usage{
GenerateNoise(X, p, type = 0)
}
\arguments{
\item{X}{an input m*n matrix}

\item{p}{the corruption probability of every pixel}

\item{type}{the type of noise, 0=white and 1=color}
}
\description{
Noise Generator
}

