% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{POPCA}
\alias{POPCA}
\title{Partially Observed PCA}
\usage{
POPCA(X, rank, p = 0)
}
\arguments{
\item{X}{an input m*n matrix}

\item{rank}{the desired rank of the output matrix}

\item{p}{the corruption probability of every pixel, if p=0 then it is PCA.}
}
\description{
Partially Observed PCA
}

