% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{RPCA}
\alias{RPCA}
\title{Robust PCA}
\usage{
RPCA(X, lambda = 1/sqrt(nrow(X)), tol = 1e-07)
}
\arguments{
\item{X}{handle to function that returns objective function values}

\item{tol}{stopping tolerance, default 1e-7}

\item{lamnda}{tuning parameter}
}
\description{
Robust PCA
}

